ATR-180: reworked and extended BQL field generation
